requirements.txt

viam-sdk==0.51.0

typing-extensions
